+++ analysis/german-powerplants.py
@@ -1,15 +1,9 @@
 import copy
 
-import cartopy.crs as ccrs
 import hvplot
 import hvplot.pandas  # noqa
 import hvplot.xarray  # noqa
-import pandas as pd
 import panel as pn
-import xarray as xr
-from holoviews import opts
-from holoviews.plotting.util import process_cmap
-from xarray import align
 
 import powerplantmatching as pm