+++ test/test_data.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 """
 Created on Wed Nov 25 08:48:04 2020
 
@@ -41,12 +40,12 @@ def test_data_request_processed(source):
 
 
 def test_OPSD_VRE():
-    df = pm.data.OPSD_VRE().empty
+    df = pm.data.OPSD_VRE()
     assert not df.empty
     assert df.Capacity.sum() > 0
 
 
-def test_OPSD_VRE():
+def test_OPSD_VRE_country():
     df = pm.data.OPSD_VRE_country("DE")
     assert not df.empty
     assert df.Capacity.sum() > 0